separating counter theme code
