+++ lib/counter/model/model.dart
@@ -1,4 +1,3 @@
-import 'package:counter2point0/counter/counter.dart';
 import 'package:counter2point0/utils/constants/constants.dart';
 import 'package:flutter/material.dart';
