+++ lib/counter/view/view.dart
@@ -6,6 +6,6 @@ import 'package:counter2point0/widgets/widgets.dart';
 
 part 'textfield_counter.dart';
 part 'buttons_row.dart';
-part 'counter_main.dart';
+part 'counter_screen.dart';
+part 'regular_counter.dart';
 part 'editable_counter.dart';
-part 'counter_number.dart';
