+++ lib/theme/controller/controller.dart
@@ -1,3 +1,5 @@
+import 'package:counter2point0/utils/constants/constants.dart';
 import 'package:flutter/material.dart';
 
 part 'theme_controller.dart';
+part 'counter_theme_contoller.dart';
